requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.28.44
botocore==1.31.44
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
crispy-bootstrap5==0.7
cryptography==41.0.3
cycler==0.11.0
defusedxml==0.7.1
Django~=4.2.4
django-allauth==0.54.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
django-htmx==1.16.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-redis==5.3.0
django-storages==1.14
django-stubs==4.2.4
drf-spectacular==0.26.4
flake8==6.1.0
fonttools==4.42.1
frozendict==2.3.8
google-api-python-client==2.98.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
iniconfig==2.0.0
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.2
multitasking==0.0.11
numpy
oauthlib==3.2.2
packaging==23.1
pandas<=2.1.0
phonenumberslite==8.13.20
Pillow==10.0.0
plotly==5.16.1
pluggy==1.2.0
pmdarima~=2.0.3
prophet~=1.1.4
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
selenium==4.12.0
setuptools==68.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
statsmodels==0.14.0
ta==0.10.2
tensorflow~=2.13.0
threadpoolctl==3.2.0
tzdata==2023.3
webencodings==0.5.1
wheel==0.41.2
whitenoise==6.5.0
yfinance==0.2.28
pip~=23.2.1
shap~=0.42.1
keras~=2.13.1
google~=3.0.0